Dockers_source_code/client_training/requirements.txt
Flask==2.2.3
grpcio==1.53.0
matplotlib==3.7.1
numpy==1.24.2
protobuf==3.20.3
PyYAML==6.0
PyYAML==6.0
Requests==2.31.0
scikit_learn==1.2.2
torch==2.0.1
torchvision==0.15.2
